Move core copy functions back to diagram
